requirements.txt
celery>=5.1.0
channels>=3.0.3
channels-redis>=3.2.0
daphne>=3.0.2
Django>=3.2.6
django-appconf>=1.0.4
django-celery-results>=2.1.0
django-cors-headers>=3.8.0
django-debug-toolbar>=3.2.1
django-filter>=2.4.0
django-redis>=5.0.0
django-simple-history>=3.0.0
django-taggit>=1.4.0
djangorestframework>=3.12.4
djangorestframework-simplejwt>=4.7.1
gunicorn>=20.1.0
mysqlclient>=2.0.3
phonenumbers>=8.12.25
Pillow>=8.2.0
pymemcache>=3.4.4
pyotp>=2.6.0
python-dateutil>=2.8.1
python-memcached>=1.59
qrcode>=7.2
redis>=3.5.3
sentry-sdk>=1.1.0
six>=1.16.0
Twisted>=21.2.0
whitenoise>=5.2.0